# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Global Justice API",
    "version": "1.0.0",
    "description": "API documentation for Global Justice Project"
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/getstate/{country_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "country_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/getcity/{state_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "state_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/get_delegate_state/{country_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "country_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/get_delegate_city/{state_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "state_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}